+++ CHANGELOG.md
@@ -7,6 +7,14 @@ release date when you use `npm version` (see `README.md`).
 
 ## [Unreleased]
 
+### Added
+
+- The ability to turn off a device in mobile
+
+- Nicer icons
+
+- Sidebar with messages is collapsible
+
 ## [0.12.1][] - 2022-06-29
 
 ### Added
@@ -255,7 +263,5 @@ release date when you use `npm version` (see `README.md`).
 [0.11.0]: https://github.com/webxdc/webxdc-dev/tree/v0.11.0
 [unreleased]: https://github.com/webxdc/webxdc-dev/compare/v0.12.0...HEAD
 [0.12.0]: https://github.com/webxdc/webxdc-dev/tree/v0.12.0
-
-
-[Unreleased]: https://github.com/webxdc/webxdc-dev/compare/v0.12.1...HEAD
+[unreleased]: https://github.com/webxdc/webxdc-dev/compare/v0.12.1...HEAD
 [0.12.1]: https://github.com/webxdc/webxdc-dev/tree/v0.12.1